requirements.txt
openai==0.25.0
python-dotenv==0.21.0
python_telegram_bot==13.15
telegram==0.0.1
